move kafka proxy config to config.go file
